[level 1] Title: 두 정수 사이의 합, Time: 11.23 ms, Memory: 79.2 MB -BaekjoonHub

+++ 프로그래머스/1/12912. 두 정수 사이의 합/두 정수 사이의 합.java
@@ -4,12 +4,12 @@ class Solution {
         if (a == b) return (long)a;
         if (a < b) {
             for (int i = a; i <= b; i++) {
-                sum += (long)i;
+                sum += i;
             }
             return sum;
         } else {
             for (int i = b; i <= a; i++) {
-                sum += (long)i;
+                sum += i;
             }
             return sum;
         }
